[rpc] Use StrView<u64> to replace the StrView<usize>

+++ crates/rooch-rpc-api/src/api/btc_api.rs
@@ -19,7 +19,7 @@ pub trait BtcAPI {
         filter: UTXOFilterView,
         // exclusive cursor if `Some`, otherwise start from the beginning
         cursor: Option<IndexerStateIDView>,
-        limit: Option<StrView<usize>>,
+        limit: Option<StrView<u64>>,
         descending_order: Option<bool>,
     ) -> RpcResult<UTXOPageView>;
 
@@ -30,7 +30,7 @@ pub trait BtcAPI {
         filter: InscriptionFilterView,
         // exclusive cursor if `Some`, otherwise start from the beginning
         cursor: Option<IndexerStateIDView>,
-        limit: Option<StrView<usize>>,
+        limit: Option<StrView<u64>>,
         descending_order: Option<bool>,
     ) -> RpcResult<InscriptionPageView>;
 }
